+++ src/Audit.WCF.Client/AuditEndpointBehavior.cs
@@ -1,5 +1,6 @@
 ﻿using System.ServiceModel.Description;
 using System.ServiceModel.Dispatcher;
+using Audit.Core;
 
 namespace Audit.Wcf.Client
 {
@@ -20,6 +21,16 @@ namespace Audit.Wcf.Client
         /// </summary>
         public bool IncludeResponseHeaders { get; set; }
 
+        /// <summary>
+        /// The factory to create the audit scopes. Default is NULL to use the globally configured AuditScopeFactory.
+        /// </summary>
+        public IAuditScopeFactory AuditScopeFactory { get; set; }
+
+        /// <summary>
+        /// The data provider to use. Default is NULL to use the globally configured AuditDataProvider.
+        /// </summary>
+        public AuditDataProvider AuditDataProvider { get; set; }
+
         public AuditEndpointBehavior()
         {
         }
@@ -36,7 +47,7 @@ namespace Audit.Wcf.Client
         }
         public void ApplyClientBehavior(ServiceEndpoint endpoint, ClientRuntime clientRuntime)
         {
-            clientRuntime.ClientMessageInspectors.Add(new AuditMessageInspector(EventType, IncludeRequestHeaders, IncludeResponseHeaders));
+            clientRuntime.ClientMessageInspectors.Add(new AuditMessageInspector(EventType, IncludeRequestHeaders, IncludeResponseHeaders, AuditScopeFactory, AuditDataProvider));
         }
         public void ApplyDispatchBehavior(ServiceEndpoint endpoint, EndpointDispatcher endpointDispatcher)
         {